requirements.txt
pypdf
Pillow
pdfplumber
